Создал отдельный подпакет для задачи и написал проверку в Example на наличие элементов в массиве
