--- src/main/java/school/faang/Main.java
@@ -1,12 +1,0 @@
-package school.faang;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Main {
-    public static void main(String[] args) {
-        int[] integers = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-
-        Example.reverseIt(integers);
-    }
-}
